requirements.txt
transformers==4.26.0
datasets==2.9.0
accelerate==0.16.0
evaluate==0.4.0
deepspeed==0.8.0
google-cloud-storage==2.7.0
google-cloud-aiplatform==1.21.0
tensorboard==2.11.2
tensorflow==2.11.0
tensorflow-io==0.27.0
tensorflow-io-gcs-filesystem==0.27.0
ninja
rouge-score 
nltk 
py7zr
